requirements.txt
streamlit
supabase
python-dotenv
pandas
numpy
plotly
openpyxl
